CXP-962: Forbid Repositories in backend, only use Queries > Connected Apps
