+++ src/Akeneo/Connectivity/Connection/back/Application/Apps/Service/CreateConnectedApp.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Connectivity\Connection\Application\Apps\Service;
 
 use Akeneo\Connectivity\Connection\Domain\Apps\Model\ConnectedApp;
-use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Repository\ConnectedAppRepositoryInterface;
+use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\CreateConnectedAppQueryInterface;
 use Akeneo\Connectivity\Connection\Domain\Marketplace\Model\App as MarketplaceApp;
 
 /**
@@ -14,11 +14,9 @@ use Akeneo\Connectivity\Connection\Domain\Marketplace\Model\App as MarketplaceAp
  */
 final class CreateConnectedApp implements CreateConnectedAppInterface
 {
-    private ConnectedAppRepositoryInterface $repository;
-
-    public function __construct(ConnectedAppRepositoryInterface $repository)
-    {
-        $this->repository = $repository;
+    public function __construct(
+        private CreateConnectedAppQueryInterface $createConnectedAppQuery,
+    ) {
     }
 
     public function execute(MarketplaceApp $marketplaceApp, array $scopes, string $connectionCode, string $userGroupName): ConnectedApp
@@ -36,7 +34,7 @@ final class CreateConnectedApp implements CreateConnectedAppInterface
             $marketplaceApp->getPartner()
         );
 
-        $this->repository->create($app);
+        $this->createConnectedAppQuery->execute($app);
 
         return $app;
     }
