+++ src/Akeneo/Connectivity/Connection/back/Domain/Apps/Persistence/Query/CreateConnectedAppQueryInterface.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query;
+
+use Akeneo\Connectivity\Connection\Domain\Apps\Model\ConnectedApp;
+
+interface CreateConnectedAppQueryInterface
+{
+    public function execute(ConnectedApp $app): void;
+}
