+++ src/Akeneo/Connectivity/Connection/back/Domain/Apps/Persistence/Query/FindAllConnectedAppsQueryInterface.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query;
+
+use Akeneo\Connectivity\Connection\Domain\Apps\Model\ConnectedApp;
+
+interface FindAllConnectedAppsQueryInterface
+{
+    /**
+     * @return array<ConnectedApp>
+     */
+    public function execute(): array;
+}
