+++ src/Akeneo/Connectivity/Connection/back/Domain/Apps/Persistence/Query/GetAccessTokenQueryInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query;
+
+/**
+ * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+interface GetAccessTokenQueryInterface
+{
+    /**
+     * @param array<string> $scopes
+     */
+    public function execute(string $clientId, array $scopes = []): ?string;
+}
