+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Persistence/Query/GetAccessTokenQuery.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\Query;
 
+use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetAccessTokenQueryInterface;
 use Doctrine\DBAL\Connection;
 
 /**
@@ -11,7 +12,7 @@ use Doctrine\DBAL\Connection;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GetAccessTokenQuery
+final class GetAccessTokenQuery implements GetAccessTokenQueryInterface
 {
     public function __construct(private Connection $connection)
     {
