+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Persistence/Query/GetAllConnectedAppsPublicIdsQuery.php
@@ -7,13 +7,14 @@ namespace Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\Query;
 use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetAllConnectedAppsPublicIdsInterface;
 use Doctrine\DBAL\Connection;
 
-class GetAllConnectedAppsPublicIdsQuery implements GetAllConnectedAppsPublicIdsInterface
+/**
+ * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+final class GetAllConnectedAppsPublicIdsQuery implements GetAllConnectedAppsPublicIdsInterface
 {
-    private Connection $connection;
-
-    public function __construct(Connection $connection)
+    public function __construct(private Connection $connection)
     {
-        $this->connection = $connection;
     }
 
     /**
