+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Persistence/Query/GetUserConsentedAuthenticationUuidQuery.php
@@ -11,13 +11,10 @@ use Doctrine\DBAL\Connection;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GetUserConsentedAuthenticationUuidQuery implements GetUserConsentedAuthenticationUuidQueryInterface
+final class GetUserConsentedAuthenticationUuidQuery implements GetUserConsentedAuthenticationUuidQueryInterface
 {
-    private Connection $connection;
-
-    public function __construct(Connection $connection)
+    public function __construct(private Connection $connection)
     {
-        $this->connection = $connection;
     }
 
     public function execute(int $userId, string $appId): string
