+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Apps/ConfirmAuthenticationEndToEnd.php
@@ -12,7 +12,7 @@ use Akeneo\Connectivity\Connection\Domain\Apps\ValueObject\ScopeList;
 use Akeneo\Connectivity\Connection\Domain\Marketplace\Model\App;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\OAuth\ClientProvider;
-use Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\DbalConnectedAppRepository;
+use Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\Query\CreateConnectedAppQuery;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\Session\AppAuthorizationSession;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\User\CreateUserGroup;
 use Akeneo\Connectivity\Connection\Infrastructure\User\Internal\CreateUser;
@@ -34,7 +34,7 @@ class ConfirmAuthenticationEndToEnd extends WebTestCase
     private PropertyAccessor $propertyAccessor;
     private ClientManagerInterface $clientManager;
     private AppAuthorizationSession $appAuthorizationSession;
-    private DbalConnectedAppRepository $repository;
+    private CreateConnectedAppQuery $createConnectedAppQuery;
     private CreateConnection $createConnection;
     private ClientProvider $clientProvider;
     private CreateUserGroup $createUserGroup;
@@ -51,7 +51,7 @@ class ConfirmAuthenticationEndToEnd extends WebTestCase
         $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
         $this->clientManager = $this->get('fos_oauth_server.client_manager.default');
         $this->appAuthorizationSession = $this->get(AppAuthorizationSession::class);
-        $this->repository = $this->get(DbalConnectedAppRepository::class);
+        $this->createConnectedAppQuery = $this->get(CreateConnectedAppQuery::class);
         $this->createConnection = $this->get(CreateConnection::class);
         $this->clientProvider = $this->get('akeneo_connectivity.connection.service.apps.client_provider');
         $this->createUserGroup = $this->get(CreateUserGroup::class);
@@ -135,7 +135,7 @@ class ConfirmAuthenticationEndToEnd extends WebTestCase
             $user->id()
         );
 
-        $this->repository->create(
+        $this->createConnectedAppQuery->execute(
             new ConnectedApp(
                 $appPublicId,
                 'App',
