+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Apps/Command/CreateAppWithAuthorizationHandlerIntegration.php
@@ -11,7 +11,7 @@ use Akeneo\Connectivity\Connection\Application\Apps\Command\RequestAppAuthorizat
 use Akeneo\Connectivity\Connection\Domain\Apps\Exception\InvalidAppAuthorizationRequestException;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
 use Akeneo\Connectivity\Connection\Domain\Settings\Persistence\Repository\ConnectionRepositoryInterface;
-use Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\DbalConnectedAppRepository;
+use Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\Query\FindOneConnectedAppByIdQuery;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\Security\ScopeMapperRegistry;
 use Akeneo\Connectivity\Connection\Tests\Integration\Mock\FakeWebMarketplaceApi;
 use Akeneo\Test\Integration\Configuration;
@@ -40,7 +40,7 @@ class CreateAppWithAuthorizationHandlerIntegration extends TestCase
     private UserManager $userManager;
     private PropertyAccessor $propertyAccessor;
     private FakeWebMarketplaceApi $webMarketplaceApi;
-    private DbalConnectedAppRepository $appRepository;
+    private FindOneConnectedAppByIdQuery $findOneConnectedAppByIdQuery;
     private ConnectionRepositoryInterface $connectionRepository;
     private RoleWithPermissionsRepository $roleWithPermissionsRepository;
     private ScopeMapperRegistry $scopeMapperRegistry;
@@ -60,7 +60,7 @@ class CreateAppWithAuthorizationHandlerIntegration extends TestCase
         $this->clientManager = $this->get('fos_oauth_server.client_manager.default');
         $this->userManager = $this->get('pim_user.manager');
         $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
-        $this->appRepository = $this->get(DbalConnectedAppRepository::class);
+        $this->findOneConnectedAppByIdQuery = $this->get(FindOneConnectedAppByIdQuery::class);
         $this->connectionRepository = $this->get('akeneo_connectivity.connection.persistence.repository.connection');
         $this->roleWithPermissionsRepository = $this->get('pim_user.repository.role_with_permissions');
         $this->scopeMapperRegistry = $this->get(ScopeMapperRegistry::class);
@@ -174,7 +174,7 @@ class CreateAppWithAuthorizationHandlerIntegration extends TestCase
 
         $this->handler->handle(new CreateAppWithAuthorizationCommand($appId));
 
-        $foundApp = $this->appRepository->findOneById($appId);
+        $foundApp = $this->findOneConnectedAppByIdQuery->execute($appId);
         Assert::assertNotNull($foundApp, 'No persisted app found');
         Assert::assertEquals($appId, $foundApp->getId());
