+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Apps/OAuth/CreateAccessTokenSpec.php
@@ -7,6 +7,7 @@ namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Apps\OAuth;
 use Akeneo\Connectivity\Connection\Application\Apps\Service\CreateAccessTokenInterface;
 use Akeneo\Connectivity\Connection\Application\RandomCodeGeneratorInterface;
 use Akeneo\Connectivity\Connection\Domain\Apps\DTO\AppConfirmation;
+use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetAccessTokenQueryInterface;
 use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetAppConfirmationQueryInterface;
 use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetConnectedAppScopesQueryInterface;
 use Akeneo\Connectivity\Connection\Domain\Apps\Persistence\Query\GetUserConsentedAuthenticationScopesQueryInterface;
@@ -15,7 +16,6 @@ use Akeneo\Connectivity\Connection\Domain\Apps\ValueObject\ScopeList;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\OAuth\ClientProviderInterface;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\OAuth\CreateAccessToken;
 use Akeneo\Connectivity\Connection\Infrastructure\Apps\OAuth\CreateJsonWebToken;
-use Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\Query\GetAccessTokenQuery;
 use Akeneo\Tool\Bundle\ApiBundle\Entity\Client;
 use Akeneo\UserManagement\Component\Model\UserInterface;
 use Akeneo\UserManagement\Component\Repository\UserRepositoryInterface;
@@ -40,7 +40,7 @@ class CreateAccessTokenSpec extends ObjectBehavior
         GetConnectedAppScopesQueryInterface $getConnectedAppScopesQuery,
         GetUserConsentedAuthenticationUuidQueryInterface $getUserConsentedAuthenticationUuidQuery,
         GetUserConsentedAuthenticationScopesQueryInterface $getUserConsentedAuthenticationScopesQuery,
-        GetAccessTokenQuery $getAccessTokenQuery,
+        GetAccessTokenQueryInterface $getAccessTokenQuery,
     ): void {
         $this->beConstructedWith(
             $storage,
@@ -74,7 +74,7 @@ class CreateAccessTokenSpec extends ObjectBehavior
         UserInterface $appUser,
         UserInterface $pimUser,
         GetUserConsentedAuthenticationScopesQueryInterface $getUserConsentedAuthenticationScopesQuery,
-        GetAccessTokenQuery $getAccessTokenQuery,
+        GetAccessTokenQueryInterface $getAccessTokenQuery,
     ): void {
         $clientProvider->findClientByAppId('client_id_1234')
             ->willReturn($client);
@@ -125,7 +125,7 @@ class CreateAccessTokenSpec extends ObjectBehavior
         GetConnectedAppScopesQueryInterface $getConnectedAppScopesQuery,
         UserInterface $pimUser,
         GetUserConsentedAuthenticationScopesQueryInterface $getUserConsentedAuthenticationScopesQuery,
-        GetAccessTokenQuery $getAccessTokenQuery,
+        GetAccessTokenQueryInterface $getAccessTokenQuery,
     ): void {
         $clientProvider->findClientByAppId('client_id_1234')
             ->willReturn($client);
@@ -263,7 +263,7 @@ class CreateAccessTokenSpec extends ObjectBehavior
         UserRepositoryInterface $userRepository,
         UserInterface $pimUser,
         GetUserConsentedAuthenticationScopesQueryInterface $getUserConsentedAuthenticationScopesQuery,
-        GetAccessTokenQuery $getAccessTokenQuery
+        GetAccessTokenQueryInterface $getAccessTokenQuery
     ): void {
         $clientProvider->findClientByAppId('client_id_1234')
             ->willReturn($client);
@@ -305,7 +305,7 @@ class CreateAccessTokenSpec extends ObjectBehavior
         UserInterface $appUser,
         UserInterface $pimUser,
         GetUserConsentedAuthenticationScopesQueryInterface $getUserConsentedAuthenticationScopesQuery,
-        GetAccessTokenQuery $getAccessTokenQuery
+        GetAccessTokenQueryInterface $getAccessTokenQuery
     ): void {
         $clientProvider->findClientByAppId('client_id_1234')
             ->willReturn($client);
